refactor: code cleanup of modals
