+++ src/App.tsx
@@ -1,10 +1,10 @@
 import { useEffect, useState } from "react";
 import { Route, Routes } from "react-router-dom";
-import InterfacesPage from "@/pages/interfaces.tsx";
-import ListenersPage from "@/pages/listeners.tsx";
+import InterfacesPage from "@/pages/interfaces/index.tsx";
+import ListenersPage from "@/pages/listeners/index.tsx";
 import LoginPage from "@/pages/login.tsx";
 import PrivateRoute from "@/pages/private.tsx";
-import AgentPage from "@/pages/agents.tsx";
+import AgentPage from "@/pages/agents/index.tsx";
 import ErrorPage from "@/pages/error.tsx";
 import { PageNotFoundError } from "@/errors/pages.ts";
 import LoadingPage from "@/pages/loading.tsx";
@@ -25,11 +25,8 @@ function App() {
         <Route element={<AgentPage />} path="/" />
         <Route element={<InterfacesPage />} path="/interfaces" />
         <Route element={<ListenersPage />} path="/listeners" />
-        <Route
-          element={<ErrorPage error={new PageNotFoundError()} />}
-          path="*"
-        />
       </Route>
+      <Route element={<ErrorPage error={new PageNotFoundError()} />} path="*" />
     </Routes>
   );
 }
